// src/Ambassador/Models/TaskList.ts
export default class TaskList {
    id: number = -1;
    program_task_id: number = -1;
    title: string = "";
    description: string = "";
    link: string = "";
    createdAt: string = "";
    task_type: number = -1;
    option: number = -1;
    updatedAt: string = "";
    custom_camp_id: number = -1;
    completion_link: string = "";
    rejected_reason: string = "";
    completion_status: number = -1;
    custom_task_id: number = -1;
    commentBody: string = "";
}